+++ crates/status/Cargo.toml
@@ -1,9 +1,7 @@
 [package]
-name = "probable-octo-computing-machine"
+name = "status"
 version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-status = { path = "./crates/status" }
-
 bevy = { git = "https://github.com/bevyengine/bevy.git" } # For now local `main` branch
